dev/kaos_dev_stream_ta_worker/src/requirements.txt
google-cloud-pubsub
google-auth
